+++ frontend/src/components/Layout.tsx
@@ -1,12 +1,13 @@
 import React from 'react';
-import { Outlet } from 'react-router-dom';
-import Header from './Header.jsx';
+import { Header } from './Header';
+import { Outlet } from 'react-router';
 
 export function Layout() {
   return (
     <React.Fragment>
       <Header />
       <main>
+        {/* @ts-expect-error we need to update react router types */}
         <Outlet />
       </main>
       <Header fixed />
