+++ frontend/src/index.jsx
@@ -6,7 +6,7 @@ import { createBrowserRouter, RouterProvider } from 'react-router-dom';
 import * as serviceWorker from './serviceWorker';
 import './i18n';
 import { UserProvider } from './context/UserContext.jsx';
-import { Layout } from './components/Layout.jsx';
+import { Layout } from './components/Layout';
 import Home from './views/home/index';
 import Profile from './views/profile';
 import Pitch from './views/pitch';
